requirements.txt
gunicorn
numpy
pillow
matplotlib
torch
torchvision
flask
opencv-python-headless

